Move Catch and add its license
